Style: 주석 추가 및 dirty code 삭제
