+++ Koco/src/main/java/icet/koco/auth/controller/AuthController.java
@@ -2,7 +2,6 @@ package icet.koco.auth.controller;
 
 import icet.koco.auth.dto.*;
 import icet.koco.auth.service.AuthService;
-import icet.koco.auth.service.LogoutService;
 import icet.koco.global.exception.UnauthorizedException;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -20,8 +19,13 @@ import org.springframework.web.bind.annotation.*;
 public class AuthController {
 
     private final AuthService authService;
-    private final LogoutService logoutService;
 
+    /**
+     * 카카오 콜백 인가 API
+     * @param code
+     * @param response
+     * @return
+     */
     @GetMapping("/callback")
     @Operation(summary = "카카오 콜백")
     public ResponseEntity<AuthResponse> kakaoCallback(@RequestParam("code") String code,
@@ -31,19 +35,30 @@ public class AuthController {
         return ResponseEntity.ok(authResponse);
     }
 
+    /**
+     * 로그아웃 API
+     * @param request
+     * @param response
+     * @return
+     */
     @Operation(summary = "로그아웃")
     @PostMapping("/logout")
     public ResponseEntity<LogoutResponse> logout(HttpServletRequest request, HttpServletResponse response) {
-        return ResponseEntity.ok(logoutService.logout(request, response));
+        return ResponseEntity.ok(authService.logout(request, response));
     }
 
+    /**
+     * 리프레쉬 토큰 기반 재발급 API
+     * @param refreshToken
+     * @param response
+     * @return
+     */
     @Operation(summary = "리프레쉬 토큰 기반 재발급")
     @PostMapping("/refresh")
     public ResponseEntity<RefreshResponse> refreshToken(
-            @CookieValue(value = "refresh_token", required = true) String refreshToken,
+            @CookieValue(value = "refresh_token") String refreshToken,
             HttpServletResponse response) {
         try {
-            System.out.println("refresh_token: " + refreshToken);
             RefreshResponse refreshResponse = authService.refreshAccessToken(refreshToken, response);
             return ResponseEntity.ok(refreshResponse);
         } catch (UnauthorizedException e) {
