--- Koco/src/main/java/icet/koco/auth/dto/KakaoCallbackRequest.java
@@ -1,11 +1,0 @@
-package icet.koco.auth.dto;
-
-import lombok.*;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class KakaoCallbackRequest {
-    private String code;
-}
