+++ Koco/src/main/java/icet/koco/auth/dto/KakaoUserResponse.java
@@ -3,41 +3,6 @@ package icet.koco.auth.dto;
 import lombok.*;
 import java.util.Optional;
 
-//@Data
-//@NoArgsConstructor
-//@AllArgsConstructor
-//@Builder
-//public class KakaoUserResponse {
-//    private Long id;
-//    private KakaoAccount kakao_account;
-//
-//    @Data
-//    @NoArgsConstructor
-//    @AllArgsConstructor
-//    @Builder
-//    public static class KakaoAccount {
-//        private String email;
-//        private String nickname;
-////        private String name;
-//    }
-//
-//    public String getEmail() {
-//        return kakao_account.getEmail();
-//    }
-//
-////    public String getName() {
-////        return kakao_account.getName();
-////    }
-//
-//    public String getName() {
-//        return kakao_account.getNickname();
-//    }
-//
-//    public String getProviderId() {
-//        return String.valueOf(id);
-//    }
-//}
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
