--- Koco/src/main/java/icet/koco/auth/dto/RefreshRequest.java
@@ -1,10 +1,0 @@
-package icet.koco.auth.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class RefreshRequest {
-    private String refreshToken;
-}
