+++ Koco/src/main/java/icet/koco/auth/repository/OAuthRepository.java
@@ -10,16 +10,10 @@ import org.springframework.transaction.annotation.Transactional;
 
 public interface OAuthRepository extends JpaRepository<OAuth, Long> {
 
-    // providerId로 OAuth 정보 조회
-    Optional<OAuth> findByProviderId(String providerId);
-
-    // provider + providerId로 정확히 식별
-    Optional<OAuth> findByProviderAndProviderId(String provider, String providerId);
-
     // user_id로 OAuth 정보 조회
     Optional<OAuth> findByUserId(Long userId);
 
-    // 사용자 ID로 refresh token을 갱신합니다.
+    // 사용자 ID로 refresh token을 갱신
     @Transactional
     @Modifying
     @Query("UPDATE OAuth o SET o.refreshToken = :refreshToken WHERE o.user.id = :userId")
