--- Koco/src/main/java/icet/koco/config/WebConfig.java
@@ -1,14 +1,0 @@
-package icet.koco.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.*;
-
-@Configuration
-public class WebConfig implements WebMvcConfigurer {
-
-    @Override
-    public void addResourceHandlers(ResourceHandlerRegistry registry) {
-        registry.addResourceHandler("/images/**")
-            .addResourceLocations("file:/Users/yeon/Desktop/Koco");
-    }
-}
