--- Koco/src/main/java/icet/koco/problemSet/repository/SurveyRepositoryCustom.java
@@ -1,8 +1,0 @@
-package icet.koco.problemSet.repository;
-
-import icet.koco.user.dto.UserCategoryStatProjection;
-import java.util.List;
-
-public interface SurveyRepositoryCustom {
-    List<UserCategoryStatProjection> calculateCorrectRateByCategory(Long userId);
-}
