+++ Koco/src/main/java/icet/koco/util/JwtTokenProvider.java
@@ -6,9 +6,12 @@ import io.jsonwebtoken.security.Keys;
 import jakarta.annotation.PostConstruct;
 import java.util.Date;
 import javax.crypto.SecretKey;
+
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
+@Slf4j
 @Component
 public class JwtTokenProvider {
 
@@ -17,15 +20,20 @@ public class JwtTokenProvider {
 
     private SecretKey key;
     private final long accessTokenValidity = 1000 * 60 * 30; // 30분
-//    private final long accessTokenValidity = 5 * 1000L; // 5초
+
 
     private final long refreshTokenValidity = 1000L * 60 * 60 * 24 * 14; // 14일
-//    private final long refreshTokenValidity = 5 * 1000L;
+
     @PostConstruct
     public void init() {
         this.key = Keys.hmacShaKeyFor(secret.getBytes());
     }
 
+    /**
+     * accessToken 생성
+     * @param user
+     * @return
+     */
     public String createAccessToken(User user) {
         return Jwts.builder()
             .setSubject(user.getId().toString())
@@ -36,6 +44,11 @@ public class JwtTokenProvider {
             .compact();
     }
 
+    /**
+     * refreshToken 생성
+     * @param user
+     * @return
+     */
     public String createRefreshToken(User user) {
         return Jwts.builder()
             .setSubject(user.getId().toString())
@@ -46,22 +59,31 @@ public class JwtTokenProvider {
     }
 
 
+    /**
+     * 토큰 검증
+     * @param token
+     * @return
+     */
+
     public boolean validateToken(String token) {
         try {
             Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(token);
             return true;
         } catch (ExpiredJwtException e) {
-            System.out.println("❌ 만료된 토큰입니다.");
-        } catch (SignatureException e) {
-            System.out.println("❌ 서명 오류 (key가 맞지 않음).");
+            log.error("Exception [Err_Msg]: {}", e.getMessage());
         } catch (MalformedJwtException e) {
-            System.out.println("❌ 잘못된 형식의 토큰.");
+            log.error("Exception [Err_Msg]: {}", e.getMessage());
         } catch (Exception e) {
-            System.out.println("❌ 기타 오류: " + e.getMessage());
+            System.out.println("기타 오류: " + e.getMessage());
         }
         return false;
     }
 
+    /**
+     * 토큰에서부터 userId 추출
+     * @param token
+     * @return
+     */
     public Long getUserIdFromToken(String token) {
         Claims claims = Jwts.parserBuilder().setSigningKey(key).build()
             .parseClaimsJws(token)
@@ -69,7 +91,11 @@ public class JwtTokenProvider {
         return Long.valueOf(claims.getSubject());
     }
 
-    // 토큰 만료시간 계산
+    /**
+     * 토큰 만료 시간 계산
+     * @param token
+     * @return
+     */
     public long getExpiration(String token) {
         return Jwts.parserBuilder()
             .setSigningKey(key)
