+++ Koco/src/main/resources/application.properties
@@ -3,9 +3,10 @@ spring.datasource.url=${DB_URL}
 spring.datasource.username=${DB_USERNAME}
 spring.datasource.password=${DB_PASSWORD}
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.show-sql=true
 spring.jpa.database=mysql
+spring.jpa.hibernate.ddl-auto=update
+#spring.jpa.show-sql=true
+
 
 
 spring.profiles.active=prod
@@ -14,7 +15,7 @@ cloud.aws.credentials.access-key=${AWS_S3_ACCESS_KEY}
 cloud.aws.credentials.secret-key=${AWS_S3_SECRET_KEY}
 
 cloud.aws.region.static=${AWS_REGION_STATIC}
-cloud.aws.s3.bucket=aposfiij
+cloud.aws.s3.bucket=${cloud.aws.s3.bucket}
 cloud.aws.stack.auto=false
 
 spring.servlet.multipart.enabled=true
